Struct strymon_communication::rpc::Outgoing

pub struct Outgoing { /* fields omitted */ }

Sender-side queue for sending out requests.

Requests can be sent to the remote peer using the Outgoing::request() method. As this method returns a future for the response, a sender can send out many concurrent outgoing requests without having to wait for responses first. This means that some requests might return before others.

Methods

impl Outgoing

pub fn request<N: Name, R: Request<N>>(&self, r: &R) -> Response<N, R>

Asynchronously sends out a request to the remote peer.

Returns a future for the pending response. The next request can be submitted without having to wait for the previous response to arrive.