Struct strymon_communication::Network

pub struct Network { /* fields omitted */ }

Handle to the networking subsystem.

Currently, this handle only stores information about the externally reachable hostname of the current process. In the future, it might be backed by a dedicated networking thread to implement its methods. Thus it recommended to only initialize it once per process.

Methods

impl Network

pub fn new<T: Into<Option<String>>>(hostname: T) -> Result<Self>

Creates a new default network handle.

The hostname argument should be set to an externally reachable hostname of the current process. If it is None, this constructor will try to retrieve it by reading the STRYMON_COMM_HOSTNAME environment variable. If both the argument and the environment variable are missing, it falls back to "localhost".

pub fn hostname(&self) -> String

Returns the configured externally reachable hostname of the current process.

impl Network

pub fn connect<E: ToSocketAddrs>(
&self,
endpoint: E
) -> Result<(Sender, Receiver)>

Connects to a socket returns two queue handles.

The endpoint socket is typically specified a (host, port) pair. The returned queue handles can be used to send and receive MessageBuf objects on that socket. Please refer to the transport module level documentation for more details.

pub fn listen<P: Into<Option<u16>>>(&self, port: P) -> Result<Listener>

Opens a new socket and returns a handle to receive incomming clients.

If the port is not specified, a random ephemerial port is chosen. Please refer to the transport module level documentation for more details.

impl Network

pub fn client<N: Name, E: ToSocketAddrs>(
&self,
endpoint: E
) -> Result<(Outgoing, Incoming<N>)>

Connects to an remote procedure call server.

Please refer to the rpc module level documentation for more details.

pub fn server<N: Name, P: Into<Option<u16>>>(
&self,
port: P
) -> Result<Server<N>>

Creates a new remote procedure call server.

If the port is not specified, a random ephemerial port is chosen. Please refer to the rpc module level documentation for more details.

impl Network

pub fn upload<P: AsRef<Path>>(&self, path: P) -> Result<Handle>

Serves the specified file on a ephemerial network port.

pub fn download<P: AsRef<Path>>(&self, url: &str, path: P) -> Result<()>

Downloads the file served at url and downloads it into path.