Struct strymon_communication::transport::Sender

pub struct Sender { /* fields omitted */ }

A queue handle to send messages on the channel.

Drop behavior:

This will close the underlying channel (including the receiving side) when dropped. In addition if the socket is still open, drop will block until the sender queue is drained.

Methods

impl Sender

pub fn send<T: Into<MessageBuf>>(&self, msg: T)

Enqueues an outgoing message.

The message might still be dropped if the underlying socket is closed by the remote receiver before the message is dequeued.